Validation for all CRUD methods

+++ src/main/java/com/example/garage/service/CarService.java
@@ -1,13 +1,14 @@
 package com.example.garage.service;
 
-import com.example.garage.dto.car.CarResponse;
-import com.example.garage.dto.common.PageResponse;
+import com.example.garage.util.Constant;
+import com.example.garage.dto.CarResponse;
+import com.example.garage.dto.PageResponse;
 import com.example.garage.entity.Car;
 import com.example.garage.exceptions.ResourceNotFoundException;
 import com.example.garage.mapper.CarMapper;
 import com.example.garage.repository.CarRepository;
-import com.example.garage.dto.car.CreateCarRequest;
-import com.example.garage.dto.car.UpdateCarRequest;
+import com.example.garage.request.CreateCarRequest;
+import com.example.garage.request.UpdateCarRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -41,14 +42,14 @@ public class CarService {
         return CarMapper.mapToDTO(car);
     }
 
-    public PageResponse getCars(String model, String color, Double price, Integer horsepower, int page, int size) {
+    public PageResponse getCars(String model, String color, Double price, Integer horsePower, int page, int size) {
         Query query = new Query();
 
         List<Criteria> filters = new ArrayList<>();
         if (model != null) filters.add(Criteria.where("model").is(model));
         if (color != null) filters.add(Criteria.where("color").is(color));
         if (price != null) filters.add(Criteria.where("price").is(price));
-        if(price != null) filters.add(Criteria.where("horsepower").gt(horsepower));
+        if(price != null) filters.add(Criteria.where("horsepower").gt(horsePower));
         if (!filters.isEmpty()) query.addCriteria(new Criteria().andOperator(filters.toArray(new Criteria[0])));
 
 
@@ -75,7 +76,7 @@ public class CarService {
 
         car.setModel(request.getModel());
         car.setColor(request.getColor());
-        car.setHorsePower(request.getHorsepower());
+        car.setHorsePower(request.getHorsePower());
         car.setPrice(request.getPrice());
 
         car = carRepository.save(car);
@@ -83,6 +84,9 @@ public class CarService {
     }
 
     public void deleteCar(String id) {
+        if (!carRepository.existsById(id)) {
+            throw new ResourceNotFoundException(Constant.CAR_NOT_FOUND);
+        }
         carRepository.deleteById(id);
     }
 
@@ -90,16 +94,16 @@ public class CarService {
     public void validation(CreateCarRequest request){
 
         if (request.getModel() == null || request.getModel().isEmpty())
-            throw new ResourceNotFoundException("model field cannot be empty");
+            throw new ResourceNotFoundException(Constant.MODEL_ERROR_MESSAGE);
 
         if (request.getColor() == null || request.getColor().isEmpty())
-            throw new ResourceNotFoundException("color field cannot be empty");
+            throw new ResourceNotFoundException(Constant.COLOR_ERROR_MESSAGE);
 
         if(request.getHorsePower() == null || request.getHorsePower() <= 0)
-            throw new ResourceNotFoundException("horsepower field cannot be empty or negative");
+            throw new ResourceNotFoundException(Constant.HP_ERROR_MESSAGE);
 
         if(request.getPrice() == null || request.getPrice() <= 0)
-            throw new ResourceNotFoundException("price field cannot be empty or negative");
+            throw new ResourceNotFoundException(Constant.PRICE_ERROR_MESSAGE);
 
     }
 }
